Extract each type to a separate file
